pyproject.toml
[tool.poetry]
name = "poetry-delta"
version = "0.1.0"
description = "Show the delta of change between commits for poetry.lock"
authors = ["gazwald <gazwald@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
GitPython = "^3.1.40"
click = "^8.1.7"
rich = "^13.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
